requirements.txt
numpy==1.21.2
tqdm==4.62.3
opencv-python==4.5.3.56
pytesseract==0.3.8
pillow==8.3.2
python-Levenshtein==0.12.2
google-cloud-speech==2.10.0
gingerit==0.8.2
nltk==3.6.5
torch==1.8.1
pytorch-crf==0.7.2
transformers==2.11.0
rank_bm25==0.2.1
Flask-Cors==3.0.10
Pathlib==1.0.1
